Add a script which spawns Xvfb, and then runs a command specified via the command-line.
This allows us to run some scripts which run X
programs (like Firefox) in a headless environment
which can be used for automation.

Xvfb is a headless X server.

I originally tried to rewrite some of the test scripts
to use the PhantomJS toolkit for doing headless testing
of web UI's, but the scripts didn't work too well with
the dynamic Dojo-based web UI.


Former-commit-id: d38f5a8e919a95bd89c5ef09c5244d71d17c874d