A whole slew of changes, unfortunately a bit intertwined.
Biggest change is for the GUI upgrade file creation.  This now
creates something that (in at least one instance) works, when
upgrading from 9.2 to 9.3.  This also required changes to
the packaging and installation scripts and files.

Unrelated to that, start adding support for trains, and let
manifest_util get a remote manifest.


Former-commit-id: 37fc299c29b30ed4a944f2fa9e046106bf51c9ab