Give Mac it's own radio button for permissions.
This is identical to unix at this point, other than we
write out a metadata file so we can keep the radio button
what it was set to.  (Without it's own metadata file it would never
persist as mac, it would reset to unix) Also the mac metadata file gets
vetoed in samba/netatalk, but since macs can speak  NFS and there's
no facility to veto files over this protocol....room for footshooting
abounds.

In the future we should switch this to be a custom dataset property
perhaps, and not use metadata files at all.  This migration for this
would involve having code that looked for metadata files and created
dataset properties, then deleted the metadata file.

I suspect due to code differentiation this won't be a straight merge
to 9.2.1-BRANCH

Ticket: #5839

Former-commit-id: 062a27776f1650ac2cd976873dce25eda87fec56