Deal with there being dead symlinks to old pools for the NFSv4 stablerestart files.
The root cause of the buggy behavior is that
os.path.exists() returns False for dead symlinks; it follows the symlink.
Because the programmer that wrote the code assumed
that os.path.exists would return True if the symlink was there
(regardless if it was valid or not) the code would try to create a symlink
where one already existed and explode with a glorious traceback.

Ticket: #6008

Former-commit-id: 636af47aa0eb42e72f6bed78c03cc07febcb10fb