Typo fix.
If you had AuthType set to Auto, because of this typo
an invalid ctl.conf would be generated and CTL would bail out.

This isn't a huge issue in master anymore since I've taken out
the ability to set the AuthType to Auto, however in 9.2.1.x
it is an issue because FreeNAS has both istgt and CTL and one can
pick between them.


Former-commit-id: a3c85dea0b26f4b24eda1a5411e42efa95b639ae