Adding the final piece in the Boot Volume Scrub 'n' Status thing
Ticket: #6195

Also fixed a bunch of bad code practices.

Reused code so that its not duplicated. (I actually had written a function
which was already ther in the monstrosity that is notifier.py, so
removing mine now and using the earlier one)

Added exception handling and throwing with suitable message in
'Scrub Now' django view.


Former-commit-id: e1e31b45894b523aa7568c04d9f416c55bcb8e19