Hook up the VMWare Plugin GUI to the autosnapper.
Ticket: #6396

With this change, if you define a VMWare Plugin Object FreeNAS
will snapshot any running VMs you have in VMWare before taking
a ZFS snapshot of the dataset or ZVol backing the VMWare datastore
the VMs are running on.

The result of this is that your ZFS snapshots will contain
coherent VM snapshots.  Useful for DR if you are replicating the ZFS
ZVols/datasets.


Former-commit-id: 4f4990a7aa3860a4685f8d264d54291ec8f3eec9