Amend login rpc call response to include session user.
Add loggedin user's name to the list of things returned as response to
the login call (in both user/pass as well as auth_token scenarios).

Add a whoami call to the SessionPlugin Provider.
(can be called using the "sessions.whoami" rpc call with no args provided to it).

Thanks to Jakub for helping me out on this.


Former-commit-id: 143c442171d32fb61f0438f9c0c005f68f1056ac