Add a function to reverse key translation for submission to middleware to GroupsStore.
If we keep the current practice of key translation for the GUI, this should probably be in FluxBase, but since that's not clear right now, I'm just gonna put it here.

If this is a terrible idea, I'm available for crucifiction 2pm to 4pm Pacific time with appointments made at least three months in advance.

Ticket: #9495
Ticket: #10254

Former-commit-id: b3b7fdc53003d5a5fad5e5261f09d6b0e0386ff3